task1-dockerize/requirements.txt
flask
streamlit
psutil
psycopg2-binary
sqlalchemy